fixed admin state

+++ routes/api/practices.js
@@ -6,11 +6,17 @@ const passport = require('passport');
 const keys = require('../../config/keys');
 //#endregion
 
+//#region Utility imports
+const acl = require('../../config/acl');
+const authorize = require('../../utils/authorize');
+//#endregion
+
 // Initialize router
 const router = express.Router();
 
 //#region Load input validation
 const isEmpty = require('../../validation/is-empty');
+//TODO: Server-side input validation.
 //#endregion
 
 // Load 'Practice' model
@@ -25,15 +31,25 @@ router.post(
   passport.authenticate('jwt', { session: false }),
   (req, res) => {
     const errors = {};
+    authorize(req, acl).then(result => {
+      if (result) {
+        const practiceFields = {};
+        if (req.body.tag) practiceFields.tag = req.body.tag;
+        if (req.body.to) practiceFields.to = req.body.to;
+        if (req.body.from) practiceFields.from = req.body.from;
+        if (req.body.information)
+          practiceFields.information = req.body.information;
 
-    const practiceFields = {};
-    if (req.body.tag) practiceFields.tag = req.body.tag;
-    if (req.body.to) practiceFields.to = req.body.to;
-    if (req.body.from) practiceFields.from = req.body.from;
-    if (req.body.information) practiceFields.information = req.body.information;
-
-    // Create
-    new Practice(practiceFields).save().then(practice => res.json(practice));
+        // Create
+        new Practice(practiceFields)
+          .save()
+          .then(practice => res.json(practice));
+      } else {
+        errors.unauthorized =
+          'Din bruker har ikke rettigheter til denne operasjonen.';
+        return res.status(401).json(errors);
+      }
+    });
   }
 );
 
@@ -44,21 +60,33 @@ router.put(
   '/:practice_id',
   passport.authenticate('jwt', { session: false }),
   (req, res) => {
-    const practiceFields = {};
-    if (req.body.tag) practiceFields.tag = req.body.tag;
-    if (req.body.to) practiceFields.to = req.body.to;
-    if (req.body.from) practiceFields.from = req.body.from;
-    if (req.body.information) practiceFields.information = req.body.information;
+    const errors = {};
+    authorize(req, acl).then(result => {
+      if (result) {
+        const practiceFields = {};
+        if (req.body.tag) practiceFields.tag = req.body.tag;
+        if (req.body.to) practiceFields.to = req.body.to;
+        if (req.body.from) practiceFields.from = req.body.from;
+        if (req.body.information)
+          practiceFields.information = req.body.information;
 
-    Practice.findOneAndUpdate(
-      { _id: req.params.practice_id },
-      { $set: practiceFields },
-      { new: true }
-    )
-      .then(practice => res.json(practice))
-      .catch(error =>
-        res.status(404).json({ nopractice: 'Fant ingen øvelse med den IDen.' })
-      );
+        Practice.findOneAndUpdate(
+          { _id: req.params.practice_id },
+          { $set: practiceFields },
+          { new: true }
+        )
+          .then(practice => res.json(practice))
+          .catch(error =>
+            res
+              .status(404)
+              .json({ nopractice: 'Fant ingen øvelse med den IDen.' })
+          );
+      } else {
+        errors.unauthorized =
+          'Din bruker har ikke rettigheter til denne operasjonen.';
+        return res.status(401).json(errors);
+      }
+    });
   }
 );
 
@@ -69,9 +97,17 @@ router.delete(
   '/:practice_id',
   passport.authenticate('jwt', { session: false }),
   (req, res) => {
-    console.log(req.params.practice_id);
-    Practice.findOneAndRemove({ _id: req.params.practice_id }).then(() => {
-      res.json({ success: true });
+    const errors = {};
+    authorize(req, acl).then(result => {
+      if (result) {
+        Practice.findOneAndRemove({ _id: req.params.practice_id }).then(() => {
+          res.json({ success: true });
+        });
+      } else {
+        errors.unauthorized =
+          'Din bruker har ikke rettigheter til denne operasjonen.';
+        return res.status(401).json(errors);
+      }
     });
   }
 );
